[needs MR_NAME] Add mrdir / mr-recache shell scripts, mrcd zsh function, and zsh completion.
This caches a directory of mr repositories in ~/.mrcache which
maps repository names to locations.  mrdir performs a lookup in
the directory by name, and mrcd will cd to a directory by name.
The cache avoids annoying delays when you have a large number
of repositories in your mr config, and can be rebuilt by running
mr-recache.